Fixes #1.
Actually, this fix creates more like inferior-lisp behavior: instead of forcing suggested text to be editable (which I think is not possible in Emacs), it will provide flexible prompt. If there is a symbol under point, it will show it as part of the prompt, marking as default value. If there is no symbol, it will show only prompt.