After successful connection, move to the end of the buffer.
In previous case, when *monroe* buffer was left running after connection was closed and has some content, putting cursor in the middle of that content and making new connection, will display prompt at cursor position. This fix forces prompt always be shown at the end of the file, after new connection was made.