Use exclusively 'clojure.repl/pst' for stacktraces.
This function is aliased to cljs.repl/pst in ClojureScript, which makes
it workable on both environments without any trickery. Previous code
worked, but ClojureScript analyzer would print warning about unknown
functions and variables.

To support older Clojure versions ('pst' is since 1.3),
monroe-old-style-stacktraces is added. As noted in docstring, it will
not work with ClojureScript.